requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0
django-cors-headers==3.6.0
django-debug-toolbar==3.1.1
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
mailjet-rest==1.3.4
mysqlclient==2.0.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2